+++ src/Rpg.Logic/Alchemy/Formulas/BaseFormulae.cs
@@ -5,10 +5,10 @@ namespace Rpg.Logic.Alchemy.Formulas
 {
     internal static class BaseFormulae
     {
-        public static double GetBase(IAlchemyEffect effect, IAlchemyOptions options)
+        public static decimal GetBase(IAlchemyEffect effect, IAlchemyOptions options)
         {
             return AlchemyConstants.IngredientMultiplier
-                * (1 + (AlchemyConstants.AlchemySkillFactor - 1) * (double)options.AlchemyLevel / 100)
+                * (1 + (AlchemyConstants.AlchemySkillFactor - 1) * (decimal)options.AlchemyLevel / 100)
                 * (1 + options.FortifyAlchemyPercent / 100);
         }
     }
