+++ src/Rpg.Models/Alchemy/Effects/FortifyCarryWeightAlchemyEffect.cs
@@ -7,9 +7,9 @@ namespace Rpg.Models.Alchemy.Effects
     {
         public FortifyCarryWeightAlchemyEffect(
             FortifyCarryWeightEffect effect,
-            double cost,
-            double duration,
-            double magnitude)
+            decimal cost,
+            decimal duration,
+            decimal magnitude)
             : base(effect, cost, duration, magnitude)
         {
         }
@@ -25,14 +25,14 @@ namespace Rpg.Models.Alchemy.Effects
 
     public static partial class AllAlchemyEffects
     {
-        internal static FortifyCarryWeightAlchemyEffect Create(this FortifyCarryWeightAlchemyEffect e, double cost, double duration, double magnitude)
+        internal static FortifyCarryWeightAlchemyEffect Create(this FortifyCarryWeightAlchemyEffect e, decimal cost, decimal duration, decimal magnitude)
         {
             return new FortifyCarryWeightAlchemyEffect(e.Effect, cost, duration, magnitude);
         }
 
         public static readonly FortifyCarryWeightAlchemyEffect FortifyCarryWeight = new FortifyCarryWeightAlchemyEffect(
             AllEffects.Restoration.FortifyCarryWeight,
-            cost: 0.15,
+            cost: 0.15M,
             duration: 300,
             magnitude: 4
         );
