+++ src/Rpg.Models/Alchemy/Effects/ParalyzeAlchemyEffect.cs
@@ -7,9 +7,9 @@ namespace Rpg.Models.Alchemy.Effects
     {
         public ParalyzeAlchemyEffect(
             ParalyzeEffect effect,
-            double cost,
-            double duration,
-            double magnitude)
+            decimal cost,
+            decimal duration,
+            decimal magnitude)
             : base(effect, cost, duration, magnitude)
         {
         }
@@ -25,7 +25,7 @@ namespace Rpg.Models.Alchemy.Effects
 
     public static partial class AllAlchemyEffects
     {
-        internal static ParalyzeAlchemyEffect Create(this ParalyzeAlchemyEffect e, double cost, double duration, double magnitude)
+        internal static ParalyzeAlchemyEffect Create(this ParalyzeAlchemyEffect e, decimal cost, decimal duration, decimal magnitude)
         {
             return new ParalyzeAlchemyEffect(e.Effect, cost, duration, magnitude);
         }
